ffmpeg: utils double check duration for mpeg ps[1/1]
PD#OTT-3599

Problem:
The end of some stream file has some error data.
So we can not find pts at the end of this file.
We can not get duration of the this file.

Solution:
Add double check for some stream file which can
not get duration.

Verify:
verify on A311

Change-Id: I52d665afe4031196682badf11c2eb3e99a14094b
Signed-off-by: Lifeng Cao <lifeng.cao@amlogic.com>